requirements.txt
# Multimodal Federated Learning Security Framework
# Python Dependencies

# Federated Learning Framework
flwr>=1.5.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning Utilities
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Progress Bars
tqdm>=4.65.0

# Experiment Tracking
tensorboard>=2.13.0

# Testing
pytest>=7.3.0
